# priv/static/swagger.json
{
  "definitions": {
...
    "FormSchemas": {
      "description": "All the form schemas",
      "items": {
        "$ref": "#/definitions/FormSchema"
      },
      "title": "FormSchemas",
      "type": "array"
    },
...
    "BatchProgram": {
      "description": "A mapping between batch and program",
      "example": {
        "batch_id": 1,
        "program_id": 1
      },
      "properties": {
        "batch_id": {
          "description": "The id of the batch",
          "type": "integer"
        },
        "program_id": {
          "description": "The id of the program",
          "type": "integer"
        }
      },
      "title": "BatchProgram",
      "type": "object"
    },
...
    "Batches": {
      "description": "All the batches",
      "items": {
        "$ref": "#/definitions/Batch"
      },
      "title": "Batches",
      "type": "array"
    },
...
    "Batch": {
      "description": "A batch in application",
      "example": {
        "contact_hours_per_week": 48,
        "name": "Delhi-12-NEET"
      },
      "properties": {
        "contact_hours_per_week": {
          "description": "Contact hours per week of a batch",
          "type": "integer"
        },
        "name": {
          "description": "The name of a batch",
          "type": "string"
        }
      },
      "title": "Batch",
      "type": "object"
    },
...
    "FormSchema": {
      "description": "A form schema in the application",
      "example": {
        "attributes": {
          "label": "First Name"
        },
        "name": "Registration"
      },
      "properties": {
        "attributes": {
          "description": "Attribute data for form schema",
          "type": "map"
        },
        "name": {
          "description": "Name of a form schema",
          "type": "string"
        }
      },
      "title": "FormSchema",
      "type": "object"
    },
...
    "BatchPrograms": {
      "description": "All batch and program mappings",
      "items": {
        "$ref": "#/definitions/BatchProgram"
      },
      "title": "BatchPrograms",
      "type": "array"
    },
...
  },
  "info": {
    "title": "DB Service application",
    "version": "1.0"
  },
  "paths": {
...
    "/api/batch-program/{batchProgramId}": {
      "delete": {
        "description": "",
        "operationId": "DbserviceWeb.BatchProgramController.delete",
        "parameters": [
          {
            "description": "The id of the batch_program record",
            "in": "path",
            "name": "batchProgramId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "",
        "tags": [
          "BatchProgram"
        ]
      },
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.BatchProgramController.show",
        "parameters": [
          {
            "description": "The id of the batch-program record",
            "in": "path",
            "name": "batchProgramId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BatchProgram"
            }
          }
        },
        "summary": "",
        "tags": [
          "BatchProgram"
        ]
      },
      "patch": {
        "description": "",
        "operationId": "DbserviceWeb.BatchProgramController.update",
        "parameters": [
          {
            "description": "The id of the batch_program record",
            "in": "path",
            "name": "batchProgramId",
            "required": true,
            "type": "integer"
          },
          {
            "description": "batch_program to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BatchProgram"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "schema": {
              "$ref": "#/definitions/BatchProgram"
            }
          }
        },
        "summary": "",
        "tags": [
          "BatchProgram"
        ]
      }
    },
...
    "/api/batch/{batchId}": {
      "delete": {
        "description": "",
        "operationId": "DbserviceWeb.BatchController.delete",
        "parameters": [
          {
            "description": "The id of the batch record",
            "in": "path",
            "name": "batchId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "",
        "tags": [
          "Batch"
        ]
      },
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.BatchController.show",
        "parameters": [
          {
            "description": "The id of the batch record",
            "in": "path",
            "name": "batchId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Batch"
            }
          }
        },
        "summary": "",
        "tags": [
          "Batch"
        ]
      },
      "patch": {
        "description": "",
        "operationId": "DbserviceWeb.BatchController.update",
        "parameters": [
          {
            "description": "The id of the batch record",
            "in": "path",
            "name": "batchId",
            "required": true,
            "type": "integer"
          },
          {
            "description": "Batch to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Batch"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "schema": {
              "$ref": "#/definitions/Batch"
            }
          }
        },
        "summary": "",
        "tags": [
          "Batch"
        ]
      }
    },
...
    "/api/form-schema/{formSchemaId}": {
      "delete": {
        "description": "",
        "operationId": "DbserviceWeb.FormSchemaController.delete",
        "parameters": [
          {
            "description": "The id of the form schema record",
            "in": "path",
            "name": "formSchemaId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "",
        "tags": [
          "FormSchema"
        ]
      },
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.FormSchemaController.show",
        "parameters": [
          {
            "description": "The id of the form schema record",
            "in": "path",
            "name": "formSchemaId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/FormSchema"
            }
          }
        },
        "summary": "",
        "tags": [
          "FormSchema"
        ]
      },
      "patch": {
        "description": "",
        "operationId": "DbserviceWeb.FormSchemaController.update",
        "parameters": [
          {
            "description": "The id of the form schema record",
            "in": "path",
            "name": "formSchemaId",
            "required": true,
            "type": "integer"
          },
          {
            "description": "Form schema to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/FormSchema"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "schema": {
              "$ref": "#/definitions/FormSchema"
            }
          }
        },
        "summary": "",
        "tags": [
          "FormSchema"
        ]
      }
    },
...
    "/api/form-schema": {
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.FormSchemaController.index",
        "parameters": [
          {
            "description": "The name of the form schema",
            "in": "query",
            "name": "name",
            "required": false,
            "type": "string"
          },
          {
            "description": "The attributes of the form schema",
            "in": "query",
            "name": "attributes",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/FormSchemas"
            }
          }
        },
        "summary": "",
        "tags": [
          "FormSchema"
        ]
      },
      "post": {
        "description": "",
        "operationId": "DbserviceWeb.FormSchemaController.create",
        "parameters": [
          {
            "description": "Form Schema to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/FormSchema"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/FormSchema"
            }
          }
        },
        "summary": "",
        "tags": [
          "FormSchema"
        ]
      }
    },
    "/api/batch-program": {
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.BatchProgramController.index",
        "parameters": [
          {
            "description": "The id the batch",
            "in": "query",
            "name": "batch_id",
            "required": false,
            "type": "string"
          },
          {
            "description": "The id the program",
            "in": "query",
            "name": "program",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BatchPrograms"
            }
          }
        },
        "summary": "",
        "tags": [
          "BatchProgram"
        ]
      },
      "post": {
        "description": "",
        "operationId": "DbserviceWeb.BatchProgramController.create",
        "parameters": [
          {
            "description": "BatchProgram to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BatchProgram"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/BatchProgram"
            }
          }
        },
        "summary": "",
        "tags": [
          "BatchProgram"
        ]
      }
    },
...
    "/api/batch": {
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.BatchController.index",
        "parameters": [
          {
            "description": "The name the batch",
            "in": "query",
            "name": "name",
            "required": false,
            "type": "string"
          },
          {
            "description": "The contact hours per week of the batch",
            "in": "query",
            "name": "contact_hours_per_week",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Batches"
            }
          }
        },
        "summary": "",
        "tags": [
          "Batch"
        ]
      },
      "post": {
        "description": "",
        "operationId": "DbserviceWeb.BatchController.create",
        "parameters": [
          {
            "description": "Batch to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Batch"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/Batch"
            }
          }
        },
        "summary": "",
        "tags": [
          "Batch"
        ]
      }
    },
...
  },
  "swagger": "2.0"
}